+++ srmt-project/src/main/java/org/cmariscal/srmt/repositories/MessageRepository.java
@@ -3,8 +3,10 @@ package org.cmariscal.srmt.repositories;
 import org.cmariscal.srmt.entities.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
+
 /**
  * Created by cmariscal on 3/5/2016.
  */
-public interface MessageRepository extends JpaRepository<Message, Long> {
+public interface MessageRepository extends JpaRepository<Message, BigInteger> {
 }
